+++ addon/SnapjawArchitect/Data/BuildInfo.lua
@@ -1,9 +1,9 @@
 -- AUTO-GENERATED FILE. DO NOT EDIT MANUALLY.
 
 SnapjawArchitectBuildInfo = {
-    generatedAt = "2026-07-21T12:06:31+00:00",
+    generatedAt = "2026-07-21T12:23:08+00:00",
     database = "tw_world",
     npcCount = 14338,
     gameObjectCount = 20953,
-    version = "1.0.0",
+    version = "1.0.1",
 }